+++ resharper/resharper-unity/src/CSharp/Daemon/CallGraph/UnityCallGraphUtil.cs
@@ -1,0 +1,20 @@
+using JetBrains.ReSharper.Psi.CSharp.Tree;
+using JetBrains.ReSharper.Psi.Tree;
+
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.CallGraph
+{
+    public class UnityCallGraphUtil
+    {
+        public static bool IsFunctionNode(ITreeNode node)
+        {
+            switch (node)
+            {
+                case IFunctionDeclaration _:
+                case ICSharpClosure _:
+                    return true;
+                default:
+                    return false;
+            }
+        }
+    }
+}
