+++ resharper/resharper-unity/src/CSharp/Daemon/Stages/BurstCodeAnalysis/CallGraph/CallGraphBurstMarksProvider.cs
@@ -6,6 +6,7 @@ using JetBrains.Diagnostics;
 using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.Daemon.CSharp.CallGraph;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.CallGraph;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.BurstCodeAnalysis.Analyzers;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.PerformanceCriticalCodeAnalysis;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.PerformanceCriticalCodeAnalysis.CallGraph;
@@ -143,7 +144,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.BurstCodeAnalys
                 if (element == StartTreeNode)
                     return true;
 
-                return !IsFunctionNode(element) && !IsBurstContextBannedNode(element);
+                return !UnityCallGraphUtil.IsFunctionNode(element) && !IsBurstContextBannedNode(element);
             }
 
             public override void ProcessBeforeInterior(ITreeNode element)
